+++ src/app/static/src/app/hintVersion.ts
@@ -1,1 +1,2 @@
-export const currentHintVersion = "1.0.1";
+export const currentHintVersion = "1.1.2";
+