+++ src/app/static/src/tests/components/plots/mapAdjustScale.test.ts
@@ -1,6 +1,6 @@
 import {mount, shallowMount} from '@vue/test-utils';
 import MapAdjustScale from "../../../app/components/plots/MapAdjustScale.vue";
-import {ColourScaleType} from "../../../app/store/plottingSelections/plottingSelections";
+import {ScaleType} from "../../../app/store/plottingSelections/plottingSelections";
 import registerTranslations from "../../../app/store/translations/registerTranslations";
 import Vuex from "vuex";
 import {emptyState} from "../../../app/root";
@@ -12,36 +12,43 @@ describe("MapAdjustScale component", () => {
     });
     registerTranslations(store);
 
+    const propsData = {
+        name: "test",
+        metadata: {
+            max: 2,
+            min: 1,
+            format: '',
+            scale: 1,
+            accuracy: null
+        },
+        show: true,
+        step: 0.1,
+        scale: {
+            type: ScaleType.Default,
+            customMin: 0,
+            customMax: 1
+        }
+    };
+
     it("does not render if show is false", () => {
         const wrapper = shallowMount(MapAdjustScale, {
-            store, propsData: {
-                show: false,
-                colourScale: {}
-            }
+            store, propsData: {...propsData, show: false}
         });
 
         expect(wrapper.findAll("div").length).toBe(0);
     });
 
+    it("renders radio button names as expected", () => {
+        const wrapper = shallowMount(MapAdjustScale, {store, propsData});
+        const name = "test-scaleType";
+        expect((wrapper.find("#type-input-default").element as HTMLInputElement).name).toBe(name);
+        expect((wrapper.find("#type-input-custom").element as HTMLInputElement).name).toBe(name);
+        expect((wrapper.find("#type-input-dynamic-full").element as HTMLInputElement).name).toBe(name);
+        expect((wrapper.find("#type-input-dynamic-filtered").element as HTMLInputElement).name).toBe(name);
+    });
+
     it("renders as expected with default scale", () => {
-        const wrapper = mount(MapAdjustScale, {
-            store, propsData: {
-                metadata: {
-                    max: 2,
-                    min: 1,
-                    format: '',
-                    scale: 1,
-                    accuracy: null
-                },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.Default,
-                    customMin: 0,
-                    customMax: 1
-                }
-            }
-        });
+        const wrapper = mount(MapAdjustScale, {store, propsData});
 
         expect((wrapper.find("#type-input-default").element as HTMLInputElement).checked).toBe(true);
         expect((wrapper.find("#type-input-custom").element as HTMLInputElement).checked).toBe(false);
@@ -53,19 +60,10 @@ describe("MapAdjustScale component", () => {
     });
 
     it("renders as expected with custom scale", () => {
-        const wrapper = mount(MapAdjustScale, {
-            store, propsData: {
-                metadata: {
-                    max: 2,
-                    min: 1,
-                    format: '',
-                    scale: 1,
-                    accuracy: null
-                },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.Custom,
+        const wrapper = mount(MapAdjustScale, {store, propsData: {
+            ...propsData,
+            scale: {
+                    type: ScaleType.Custom,
                     customMin: 0,
                     customMax: 1
                 }
@@ -93,17 +91,9 @@ describe("MapAdjustScale component", () => {
     it("renders as expected with full dynamic scale", () => {
         const wrapper = mount(MapAdjustScale, {
             store, propsData: {
-                metadata: {
-                    max: 2,
-                    min: 1,
-                    format: '',
-                    scale: 1,
-                    accuracy: null
-                },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.DynamicFull,
+                ...propsData,
+                scale: {
+                    type: ScaleType.DynamicFull,
                     customMin: 0,
                     customMax: 1
                 }
@@ -122,6 +112,7 @@ describe("MapAdjustScale component", () => {
     it("renders scale input modifier when scale parameter given", () => {
         const wrapper = mount(MapAdjustScale, {
             store, propsData: {
+                ...propsData,
                 metadata: {
                     max: 2,
                     min: 1,
@@ -129,10 +120,8 @@ describe("MapAdjustScale component", () => {
                     scale: 100,
                     accuracy: null
                 },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.DynamicFull,
+                scale: {
+                    type: ScaleType.DynamicFull,
                     customMin: 2,
                     customMax: 3
                 }
@@ -146,17 +135,9 @@ describe("MapAdjustScale component", () => {
     it("renders as expected with filtered dynamic scale", () => {
         const wrapper = mount(MapAdjustScale, {
             store, propsData: {
-                metadata: {
-                    max: 2,
-                    min: 1,
-                    format: '',
-                    scale: 1,
-                    accuracy: null
-                },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.DynamicFiltered,
+                ...propsData,
+                scale: {
+                    type: ScaleType.DynamicFiltered,
                     customMin: 0,
                     customMax: 1
                 }
@@ -172,31 +153,50 @@ describe("MapAdjustScale component", () => {
         expect((wrapper.find("#custom-max-input").element as HTMLInputElement).disabled).toBe(true);
     });
 
+    it("renders as expected when hide props both false", () => {
+        const wrapper = mount(MapAdjustScale, {store, propsData});
+
+        expect(wrapper.find(".static-container").exists()).toBe(true);
+        expect(wrapper.find(".static-default").exists()).toBe(true);
+        expect(wrapper.find(".static-custom").exists()).toBe(true);
+        expect(wrapper.find(".static-custom-values").exists()).toBe(true);
+    });
+
+    it("renders as expected when hide static defaults", () => {
+        const wrapper = mount(MapAdjustScale, {store, propsData: {...propsData, hideStaticDefault: true}});
+
+        expect(wrapper.find(".static-container").exists()).toBe(true);
+        expect(wrapper.find(".static-default").exists()).toBe(false);
+        expect(wrapper.find(".static-custom").exists()).toBe(true);
+        expect(wrapper.find(".static-custom-values").exists()).toBe(true);
+    });
+
+    it("renders as expected when hide static custom", () => {
+        const wrapper = mount(MapAdjustScale, {store, propsData: {...propsData, hideStaticCustom: true}});
+
+        expect(wrapper.find(".static-container").exists()).toBe(true);
+        expect(wrapper.find(".static-default").exists()).toBe(true);
+        expect(wrapper.find(".static-custom").exists()).toBe(false);
+        expect(wrapper.find(".static-custom-values").exists()).toBe(false);
+    });
+
+    it("renders as expected when hide static default and custom", () => {
+        const wrapper = mount(MapAdjustScale, {store, propsData: {...propsData, hideStaticDefault: true, hideStaticCustom: true}});
+
+        expect(wrapper.find(".static-container").exists()).toBe(false);
+        expect(wrapper.find(".static-default").exists()).toBe(false);
+        expect(wrapper.find(".static-custom").exists()).toBe(false);
+        expect(wrapper.find(".static-custom-values").exists()).toBe(false);
+    });
+
     it("emits update event when type changes", () => {
-        const wrapper = mount(MapAdjustScale, {
-            store, propsData: {
-                metadata: {
-                    max: 2,
-                    min: 1,
-                    format: '',
-                    scale: 1,
-                    accuracy: null
-                },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.Default,
-                    customMin: 0,
-                    customMax: 1
-                }
-            }
-        });
+        const wrapper = mount(MapAdjustScale, {store, propsData});
 
         wrapper.find("#type-input-custom").trigger("click");
 
         expect(wrapper.emitted("update").length).toBe(1);
         expect(wrapper.emitted("update")[0][0]).toStrictEqual({
-            type: ColourScaleType.Custom,
+            type: ScaleType.Custom,
             customMin: 0,
             customMax: 1
         });
@@ -205,7 +205,7 @@ describe("MapAdjustScale component", () => {
 
         expect(wrapper.emitted("update").length).toBe(2);
         expect(wrapper.emitted("update")[1][0]).toStrictEqual({
-            type: ColourScaleType.Default,
+            type: ScaleType.Default,
             customMin: 0,
             customMax: 1
         });
@@ -214,7 +214,7 @@ describe("MapAdjustScale component", () => {
 
         expect(wrapper.emitted("update").length).toBe(3);
         expect(wrapper.emitted("update")[1][0]).toStrictEqual({
-            type: ColourScaleType.DynamicFull,
+            type: ScaleType.DynamicFull,
             customMin: 0,
             customMax: 1
         });
@@ -223,7 +223,7 @@ describe("MapAdjustScale component", () => {
 
         expect(wrapper.emitted("update").length).toBe(4);
         expect(wrapper.emitted("update")[1][0]).toStrictEqual({
-            type: ColourScaleType.DynamicFiltered,
+            type: ScaleType.DynamicFiltered,
             customMin: 0,
             customMax: 1
         });
@@ -232,17 +232,9 @@ describe("MapAdjustScale component", () => {
     it("emits update event when custom min or max changes", () => {
         const wrapper = mount(MapAdjustScale, {
             store, propsData: {
-                metadata: {
-                    max: 2,
-                    min: 1,
-                    format: '',
-                    scale: 1,
-                    accuracy: null
-                },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.Custom,
+                ...propsData,
+                scale: {
+                    type: ScaleType.Custom,
                     customMin: 0,
                     customMax: 1
                 }
@@ -253,7 +245,7 @@ describe("MapAdjustScale component", () => {
         wrapper.find("#custom-min-input").trigger("change");
         expect(wrapper.emitted("update").length).toBe(1);
         expect(wrapper.emitted("update")[0][0]).toStrictEqual({
-            type: ColourScaleType.Custom,
+            type: ScaleType.Custom,
             customMin: 0.5,
             customMax: 1
         });
@@ -262,7 +254,7 @@ describe("MapAdjustScale component", () => {
         wrapper.find("#custom-max-input").trigger("keyup");
         expect(wrapper.emitted("update").length).toBe(2);
         expect(wrapper.emitted("update")[1][0]).toStrictEqual({
-            type: ColourScaleType.Custom,
+            type: ScaleType.Custom,
             customMin: 0.5,
             customMax: 1.5
         });
@@ -271,17 +263,9 @@ describe("MapAdjustScale component", () => {
     it("does not emit update event when type is Custom and custom max is not greater than custom min", () => {
         const wrapper = mount(MapAdjustScale, {
             store, propsData: {
-                metadata: {
-                    max: 2,
-                    min: 1,
-                    format: '',
-                    scale: 1,
-                    accuracy: null
-                },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.Custom,
+                ...propsData,
+                scale: {
+                    type: ScaleType.Custom,
                     customMin: 0,
                     customMax: 1
                 }
@@ -297,9 +281,11 @@ describe("MapAdjustScale component", () => {
         expect(wrapper.emitted("update").length).toBe(1);
     });
 
-    it("updates colourScaleToAdjust when colourScale property changes", () => {
+    it("updates scaleToAdjust when scale property changes", () => {
         const wrapper = mount(MapAdjustScale, {
+            name: "test",
             store, propsData: {
+                ...propsData,
                 metadata: {
                     max: 2,
                     min: 1,
@@ -307,10 +293,8 @@ describe("MapAdjustScale component", () => {
                     scale: 1,
                     accuracy: null
                 },
-                show: true,
-                step: 0.1,
-                colourScale: {
-                    type: ColourScaleType.Custom,
+                scale: {
+                    type: ScaleType.Custom,
                     customMin: 0,
                     customMax: 1
                 }
@@ -318,13 +302,12 @@ describe("MapAdjustScale component", () => {
         });
 
         const newScale = {
-            type: ColourScaleType.Default,
+            type: ScaleType.Default,
             customMin: 1,
             customMax: 2
         };
-        wrapper.setProps({colourScale: newScale});
+        wrapper.setProps({scale: newScale});
 
-        expect((wrapper.vm as any).colourScaleToAdjust).toStrictEqual(newScale);
-
+        expect((wrapper.vm as any).scaleToAdjust).toStrictEqual(newScale);
     });
 });